track last song play timestamp random song picks out of last 100 songs not bing played changed !current a little

+++ src/main/java/novaz/db/model/OMusic.java
@@ -6,6 +6,6 @@ public class OMusic {
 	public String filename = "";
 	public String title = "";
 	public String artist = "";
-	public int lastplaydate = 0;
+	public long lastplaydate = 0;
 	public int banned = 0;
 }
